+++ src/plugins/plugin_single.c
@@ -20,7 +20,7 @@
  *  along with this program; if not, write to the Free Software
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: plugin_single.c,v 1.4 2010-03-09 10:00:14 Isibaar Exp $
+ * $Id: plugin_single.c,v 1.4.2.1 2010-12-28 19:19:57 Isibaar Exp $
  *
  ****************************************************************************/
 
@@ -187,8 +187,8 @@ rc_single_after(rc_single_t * rc,
 	rc->total_size += data->length;
 
 	/* Compute the deviation from expected total size */
-	deviation = 
-		rc->total_size - rc->bytes_per_sec * rc->time;
+	deviation = (int64_t)
+		(rc->total_size - rc->bytes_per_sec * rc->time);
 
 	averaging_period = (double) rc->averaging_period;
